Olio-oppitunnin toinen osa

+++ src/viikko5/oliot/kaupungit_v1/Kaupunki.java
@@ -6,8 +6,7 @@ public class Kaupunki {
     int vakiluku;
     double pintaAla;
 
-    public int laskeVaestontiheys() {
-        // this-muuttuja viittaa aina siihen olioon, jonka metodia suoritetaan
-        return (int) (this.vakiluku / this.pintaAla);
+    public double laskeVaestontiheys() {
+        return this.vakiluku / this.pintaAla;
     }
 }
